# contracts/openapi/foyer-v1.yaml
openapi: 3.1.0
info:
  title: Foyer Server API
  version: 0.4.0
  description: |
    Contract for the self-hosted Foyer Rust service.

    This version covers notes, bookmarks, tasks, contacts, calendar, and device-key
    authentication. Clients communicate only through this contract and
    `contracts/auth/v1/`. They do not receive PostgreSQL credentials, DAV
    credentials, raw iCalendar/vCard payloads, or operation/checkpoint tables.

    Calendar, tasks, and contacts are normalized projections of Radicale. Notes and
    bookmarks are PostgreSQL-canonical. All mutations use client-generated operation
    identifiers. DAV-backed writes also carry the last observed ETag or revision.

    The deleted Cloudflare/Flue HTTP API is not a compatibility target.

    ## Authentication

    Production identity is a manually enrolled ECDSA P-256 device key (ADR 0023).

    - `POST /v1/auth/challenges` with `{ "deviceKeyId": "<thumbprint>" }`.
    - Sign the opaque `signingPayload` bytes with ECDSA P-256/SHA-256.
    - `POST /v1/auth/sessions` with `{ "challengeId", "signature" }` where the
      signature is IEEE P1363 `r||s`, base64url without padding.
    - Send `Authorization: Bearer <accessToken>` on every authenticated route.
    - `GET /v1/auth/jwks` publishes the server ES256 verification key.
    - `GET /v1/sync/credentials` issues a five-minute PowerSync JWT with a
      distinct audience. PowerSync verifies only the public JWKS.

...

...
paths:
...
  /v1/sync/credentials:
    get:
      operationId: getSyncCredentials
      summary: Issue a short-lived PowerSync JWT
      description: |
        Returns a five-minute ES256 JWT and the client-facing PowerSync URL. The
        token uses a distinct audience, is scoped to the current user and device,
        and is not a database credential. PowerSync verifies it with
        `GET /v1/auth/jwks`.
      responses:
        "200":
          description: Credentials the official PowerSync client SDKs can present.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncCredentials"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/folders:
    get:
      operationId: listFolders
      summary: List the current user's live folders
      responses:
        "200":
          description: Folders in deterministic parent, position, name, id order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createFolder
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "200":
          description: The folder was created or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
...
  /v1/folders/{folderId}/rename:
    post:
      operationId: renameFolder
      summary: Rename a folder
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFolderRequest"
      responses:
        "200":
          description: The folder was renamed or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/folders/{folderId}/move:
    post:
      operationId: moveFolder
      summary: Move a folder or change its position
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveFolderRequest"
      responses:
        "200":
          description: The folder was moved or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/folders/{folderId}/delete:
    post:
      operationId: deleteFolder
      summary: Tombstone an empty folder
      parameters:
        - $ref: "#/components/parameters/folderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          description: The folder was tombstoned or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
  /v1/notes:
    get:
      operationId: listNotes
      summary: List the current user's live notes
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notes in deterministic updated-at, id order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createNote
      summary: Create a Markdown note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "200":
          description: The note was created or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
        "422":
          $ref: "#/components/responses/SemanticRejection"
...
  /v1/notes/{noteId}/update:
    post:
      operationId: updateNote
      summary: Replace a note's title and Markdown body
      parameters:
        - $ref: "#/components/parameters/noteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: The note was updated or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/notes/{noteId}/move:
    post:
      operationId: moveNote
      summary: Move a note into another folder
      parameters:
        - $ref: "#/components/parameters/noteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveNoteRequest"
      responses:
        "200":
          description: The note was moved or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
  /v1/notes/{noteId}/delete:
    post:
      operationId: deleteNote
      summary: Tombstone a note
      parameters:
        - $ref: "#/components/parameters/noteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          description: The note was tombstoned or the operation was replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"
...
components:
  securitySchemes:
    foyerAccess:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Five-minute ES256 access JWT issued by POST /v1/auth/sessions. Claims
        are documented in contracts/auth/v1/.
    developmentBearer:
      type: http
      scheme: bearer
      description: >
        Development-only static token from FOYER_DEV_TOKEN. Rejected unless the
        process is running with FOYER_SERVER_ENV=development.
  parameters:
    folderId:
      name: folderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    noteId:
      name: noteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Forbidden:
      description: The token is not usable in this process environment.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: The entity does not exist for this user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Conflict:
      description: Stale expected revision or identifier collision.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Gone:
      description: The entity is tombstoned and cannot be resurrected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidRequest:
      description: The request failed input bounds or schema validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    SemanticRejection:
      description: >
        The mutation is well-formed but violates folder emptiness, parent, cycle,
        or limit rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    SyncCredentials:
      type: object
      additionalProperties: false
      required:
        - endpoint
        - token
        - expiresAt
        - userId
      properties:
        endpoint:
          type: string
          format: uri
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: string
...
    ErrorBody:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - forbidden
            - not_found
            - gone
            - stale_revision
            - conflict
            - invalid_request
            - invalid_parent
            - cycle
            - folder_not_empty
            - address_book_not_empty
            - stale_etag
            - limit_exceeded
            - unavailable
            - rate_limited
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Folder:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - parentId
        - name
        - position
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        parentId:
          type:
            - string
            - "null"
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 80
        position:
          type: integer
          minimum: 0
          maximum: 100000
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    FolderList:
      type: object
      additionalProperties: false
      required:
        - folders
      properties:
        folders:
          type: array
          items:
            $ref: "#/components/schemas/Folder"
    Note:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - folderId
        - title
        - body
        - revision
        - createdAt
        - updatedAt
        - deletedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        folderId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          maxLength: 262144
          description: Lossless Markdown source. The server never executes HTML.
        revision:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    NoteList:
      type: object
      additionalProperties: false
      required:
        - notes
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
    CreateFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - name
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 80
        parentId:
          type:
            - string
            - "null"
          format: uuid
        position:
          type: integer
          minimum: 0
          maximum: 100000
    RenameFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - name
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 80
    MoveFolderRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        parentId:
          type:
            - string
            - "null"
          format: uuid
        position:
          type: integer
          minimum: 0
          maximum: 100000
    CreateNoteRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - id
        - folderId
        - title
        - body
      properties:
        operationId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          maxLength: 262144
    UpdateNoteRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - title
        - body
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          maxLength: 262144
    MoveNoteRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
        - folderId
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
        folderId:
          type: string
          format: uuid
    DeleteRequest:
      type: object
      additionalProperties: false
      required:
        - operationId
        - expectedRevision
      properties:
        operationId:
          type: string
          format: uuid
        expectedRevision:
          type: integer
          minimum: 1
...
security:
  - foyerAccess: []
  - developmentBearer: []